refactor(docs): autocomplete dx
