+++ apps/docs/content/components/autocomplete/custom-filtering.ts
@@ -1,95 +1,8 @@
-const data = `export const animals = [
-  {label: "Cat", value: "cat", description: "The second most popular pet in the world"},
-  {label: "Dog", value: "dog", description: "The most popular pet in the world"},
-  {label: "Elephant", value: "elephant", description: "The largest land animal"},
-  {label: "Lion", value: "lion", description: "The king of the jungle"},
-  {label: "Tiger", value: "tiger", description: "The largest cat species"},
-  {label: "Giraffe", value: "giraffe", description: "The tallest land animal"},
-  {
-    label: "Dolphin",
-    value: "dolphin",
-    description: "A widely distributed and diverse group of aquatic mammals",
-  },
-  {label: "Penguin", value: "penguin", description: "A group of aquatic flightless birds"},
-  {label: "Zebra", value: "zebra", description: "A several species of African equids"},
-  {
-    label: "Shark",
-    value: "shark",
-    description: "A group of elasmobranch fish characterized by a cartilaginous skeleton",
-  },
-  {
-    label: "Whale",
-    value: "whale",
-    description: "Diverse group of fully aquatic placental marine mammals",
-  },
-  {label: "Otter", value: "otter", description: "A carnivorous mammal in the subfamily Lutrinae"},
-  {label: "Crocodile", value: "crocodile", description: "A large semiaquatic reptile"},
-];`;
-
-const App = `import {Autocomplete, AutocompleteItem} from "@nextui-org/react";
-import {animals} from "./data";
-
-export default function App() {
-  const myFilter = (textValue, inputValue) => {
-    if (inputValue.length === 0) {
-      return true;
-    }
-
-    // Normalize both strings so we can slice safely
-    // take into account the ignorePunctuation option as well...
-    textValue = textValue.normalize("NFC").toLocaleLowerCase();
-    inputValue = inputValue.normalize("NFC").toLocaleLowerCase();
-
-    return textValue.slice(0, inputValue.length) === inputValue;
-  };
-
-  return (
-    <Autocomplete
-      allowsCustomValue
-      className="max-w-xs"
-      defaultFilter={myFilter}
-      defaultItems={animals}
-      label="Search an animal"
-      variant="bordered"
-    >
-      {(item) => <AutocompleteItem key={item.value}>{item.label}</AutocompleteItem>}
-    </Autocomplete>
-  );
-}`;
-
-const AppTs = `import {Autocomplete, AutocompleteItem} from "@nextui-org/react";
-import {animals} from "./data";
-
-export default function App() {
-  const myFilter = (textValue: string, inputValue: string) => {
-    if (inputValue.length === 0) {
-      return true;
-    }
-
-    // Normalize both strings so we can slice safely
-    // take into account the ignorePunctuation option as well...
-    textValue = textValue.normalize("NFC").toLocaleLowerCase();
-    inputValue = inputValue.normalize("NFC").toLocaleLowerCase();
-
-    return textValue.slice(0, inputValue.length) === inputValue;
-  };
-
-  return (
-    <Autocomplete
-      className="max-w-xs"
-      defaultFilter={myFilter}
-      defaultItems={animals}
-      label="Search an animal"
-      variant="bordered"
-    >
-      {(item) => <AutocompleteItem key={item.value}>{item.label}</AutocompleteItem>}
-    </Autocomplete>
-  );
-}`;
+import App from "./custom-filtering.raw.jsx?raw";
+import AppTs from "./custom-filtering.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
-  "/data.js": data,
 };
 
 const reactTs = {
