+++ apps/docs/content/components/autocomplete/fully-controlled.ts
@@ -1,187 +1,11 @@
-const data = `export const animals = [
-  {label: "Cat", value: "cat", description: "The second most popular pet in the world"},
-  {label: "Dog", value: "dog", description: "The most popular pet in the world"},
-  {label: "Elephant", value: "elephant", description: "The largest land animal"},
-  {label: "Lion", value: "lion", description: "The king of the jungle"},
-  {label: "Tiger", value: "tiger", description: "The largest cat species"},
-  {label: "Giraffe", value: "giraffe", description: "The tallest land animal"},
-  {
-    label: "Dolphin",
-    value: "dolphin",
-    description: "A widely distributed and diverse group of aquatic mammals",
-  },
-  {label: "Penguin", value: "penguin", description: "A group of aquatic flightless birds"},
-  {label: "Zebra", value: "zebra", description: "A several species of African equids"},
-  {
-    label: "Shark",
-    value: "shark",
-    description: "A group of elasmobranch fish characterized by a cartilaginous skeleton",
-  },
-  {
-    label: "Whale",
-    value: "whale",
-    description: "Diverse group of fully aquatic placental marine mammals",
-  },
-  {label: "Otter", value: "otter", description: "A carnivorous mammal in the subfamily Lutrinae"},
-  {label: "Crocodile", value: "crocodile", description: "A large semiaquatic reptile"},
-];`;
-
-const App = `import {Autocomplete, AutocompleteItem} from "@nextui-org/react";
-import {useFilter} from "@react-aria/i18n";
-import {animals} from "./data";
-
-export default function App() {
-  // Store Autocomplete input value, selected option, open state, and items
-  // in a state tracker
-  const [fieldState, setFieldState] = React.useState({
-    selectedKey: "",
-    inputValue: "",
-    items: animals,
-  });
-
-  // Implement custom filtering logic and control what items are
-  // available to the Autocomplete.
-  const {startsWith} = useFilter({sensitivity: "base"});
-
-  // Specify how each of the Autocomplete values should change when an
-  // option is selected from the list box
-  const onSelectionChange = (key) => {
-    setFieldState((prevState) => {
-      let selectedItem = prevState.items.find((option) => option.value === key);
-
-      return {
-        inputValue: selectedItem?.label || "",
-        selectedKey: key,
-        items: animals.filter((item) => startsWith(item.label, selectedItem?.label || "")),
-      };
-    });
-  };
-
-  // Specify how each of the Autocomplete values should change when the input
-  // field is altered by the user
-  const onInputChange = (value) => {
-    setFieldState((prevState) => ({
-      inputValue: value,
-      selectedKey: value === "" ? null : prevState.selectedKey,
-      items: animals.filter((item) => startsWith(item.label, value)),
-    }));
-  };
-
-  // Show entire list if user opens the menu manually
-  const onOpenChange = (isOpen, menuTrigger) => {
-    if (menuTrigger === "manual" && isOpen) {
-      setFieldState((prevState) => ({
-        inputValue: prevState.inputValue,
-        selectedKey: prevState.selectedKey,
-        items: animals,
-      }));
-    }
-  };
-
-  return (
-    <Autocomplete
-      className="max-w-xs"
-      inputValue={fieldState.inputValue}
-      items={fieldState.items}
-      label="Favorite Animal"
-      placeholder="Search an animal"
-      selectedKey={fieldState.selectedKey}
-      variant="bordered"
-      onInputChange={onInputChange}
-      onOpenChange={onOpenChange}
-      onSelectionChange={onSelectionChange}
-    >
-      {(item) => <AutocompleteItem key={item.value}>{item.label}</AutocompleteItem>}
-    </Autocomplete>
-  );
-}`;
-
-const AppTs = `import {Autocomplete, AutocompleteItem, MenuTriggerAction} from "@nextui-org/react";
-import {useFilter} from "@react-aria/i18n";
-import {animals} from "./data";
-
-type FieldState = {
-  selectedKey: React.Key | null;
-  inputValue: string;
-  items: typeof animals;
-};
-
-export default function App() {
-  // Store Autocomplete input value, selected option, open state, and items
-  // in a state tracker
-  const [fieldState, setFieldState] = React.useState<FieldState>({
-    selectedKey: "",
-    inputValue: "",
-    items: animals,
-  });
-
-  // Implement custom filtering logic and control what items are
-  // available to the Autocomplete.
-  const {startsWith} = useFilter({sensitivity: "base"});
-
-  // Specify how each of the Autocomplete values should change when an
-  // option is selected from the list box
-  const onSelectionChange = (key: React.Key) => {
-    setFieldState((prevState) => {
-      let selectedItem = prevState.items.find((option) => option.value === key);
-
-      return {
-        inputValue: selectedItem?.label || "",
-        selectedKey: key,
-        items: animals.filter((item) => startsWith(item.label, selectedItem?.label || "")),
-      };
-    });
-  };
-
-  // Specify how each of the Autocomplete values should change when the input
-  // field is altered by the user
-  const onInputChange = (value: string) => {
-    setFieldState((prevState) => ({
-      inputValue: value,
-      selectedKey: value === "" ? null : prevState.selectedKey,
-      items: animals.filter((item) => startsWith(item.label, value)),
-    }));
-  };
-
-  // Show entire list if user opens the menu manually
-  const onOpenChange = (isOpen: boolean, menuTrigger: MenuTriggerAction) => {
-    if (menuTrigger === "manual" && isOpen) {
-      setFieldState((prevState) => ({
-        inputValue: prevState.inputValue,
-        selectedKey: prevState.selectedKey,
-        items: animals,
-      }));
-    }
-  };
-
-  return (
-    <Autocomplete
-      className="max-w-xs"
-      inputValue={fieldState.inputValue}
-      items={fieldState.items}
-      label="Favorite Animal"
-      placeholder="Search an animal"
-      selectedKey={fieldState.selectedKey}
-      variant="bordered"
-      onInputChange={onInputChange}
-      onOpenChange={onOpenChange}
-      onSelectionChange={onSelectionChange}
-    >
-      {(item) => <AutocompleteItem key={item.value}>{item.label}</AutocompleteItem>}
-    </Autocomplete>
-  );
-}`;
+import App from "./fully-controlled.raw.tsx?raw";
+import AppTs from "./fully-controlled.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
-  "/data.js": data,
-};
-
-const reactTs = {
   "/App.tsx": AppTs,
 };
 
 export default {
   ...react,
-  ...reactTs,
 };
