+++ apps/docs/content/components/autocomplete/sections.ts
@@ -1,35 +1,4 @@
-const App = `import {Autocomplete, AutocompleteItem, AutocompleteSection} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Autocomplete
-      label="Favorite Animal"
-      placeholder="Search an animal"
-      className="max-w-xs"
-    >
-      <AutocompleteSection showDivider title="Mammals">
-        <AutocompleteItem key="Lion">Lion</AutocompleteItem>
-        <AutocompleteItem key="Tiger">Tiger</AutocompleteItem>
-        <AutocompleteItem key="Elephant">Elephant</AutocompleteItem>
-        <AutocompleteItem key="Kangaroo">Kangaroo</AutocompleteItem>
-        <AutocompleteItem key="Panda">Panda</AutocompleteItem>
-        <AutocompleteItem key="Giraffe">Giraffe</AutocompleteItem>
-        <AutocompleteItem key="Zebra">Zebra</AutocompleteItem>
-        <AutocompleteItem key="Cheetah">Cheetah</AutocompleteItem>
-      </AutocompleteSection>
-      <AutocompleteSection title="Birds">
-        <AutocompleteItem key="Eagle">Eagle</AutocompleteItem>
-        <AutocompleteItem key="Parrot">Parrot</AutocompleteItem>
-        <AutocompleteItem key="Penguin">Penguin</AutocompleteItem>
-        <AutocompleteItem key="Ostrich">Ostrich</AutocompleteItem>
-        <AutocompleteItem key="Peacock">Peacock</AutocompleteItem>
-        <AutocompleteItem key="Swan">Swan</AutocompleteItem>
-        <AutocompleteItem key="Falcon">Falcon</AutocompleteItem>
-        <AutocompleteItem key="Flamingo">Flamingo</AutocompleteItem>
-      </AutocompleteSection>
-    </Autocomplete>
-  );
-}`;
+import App from "./sections.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
