+++ apps/docs/content/docs/components/dropdown.mdx
@@ -42,18 +42,18 @@ NextUI exports 5 dropdown-related components:
 <ImportTabs
   commands={{
     main: `import {
-  Dropdown,
-  DropdownTrigger,
-  DropdownMenu,
-  DropdownSection,
-  DropdownItem
+    Dropdown,
+    DropdownTrigger,
+    DropdownMenu,
+    DropdownSection,
+    DropdownItem
 } from "@nextui-org/react";`,
     individual: `import {
-  Dropdown,
-  DropdownTrigger,
-  DropdownMenu,
-  DropdownSection,
-  DropdownItem
+    Dropdown,
+    DropdownTrigger,
+    DropdownMenu,
+    DropdownSection,
+    DropdownItem
 } from "@nextui-org/dropdown";`,
   }}
 />
